requirements.txt
python-Levenshtein==0.25.1
twitchio==2.10.0
howlongtobeatpy==1.0.18
python-steam-api==2.0.4
google-api-python-client==2.144.0
gspread==6.1.2
requests==2.32.3
python-telegram-bot==20.7